Rename query-serialization -> query-history-serializer
