+++ extensions/ql-vscode/src/query-history-serializater.ts
